Update with some changes from WABS system
Planning on having another crack at this I wanted to bring the repo more inline with where [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) is at.

So, we have brought in `.editorconfig`, and the switch to ESlint for linting (though not currently working). We also updated how the database is cleaned prior to running the test suite to match what we now do.

We also now seed the test DB in system, which I wanted to incorporate. It was easier to just bring all current models across and convert them to ESM than try to hand pick stuff. This also then meant migrating all model helpers, and migrations!