Add first example reference model test (Regions)
Now we've imported and updated all our models it is time to start adding some example tests to drive research of node-test as an alternate to Lab further.

In database terms, all models are accessed and used in the same way via Objection.js. However, we split them into two camps for testing; reference models and transaction models.

Reference models are things we don't expect to be creating within the service as part of daily use. They are records that would be seeded, for example, regions.

Transaction models are things we expect to be created, for example, bill runs or licences.

When it comes to testing we seed the tables that back the reference models, hence their helpers generally have a `select()` method. Transaction models, when needed, will be created during the testing. Their helpers generally have an `add()` method.

When we first tried this test, node-test wouldn't quit. It wouldn't even move to the next spec after `region.model.test.js`. This is a problem we first faced when pulling this project together. Time away has clearly helped because we were able to diagnose the cause as the database connection we create blocks node test from exiting (see [test_runner: allow to force exit after all tests finished](https://github.com/nodejs/node/issues/49925) for a deeper dive into the issue which gave us our a-ha! moment).

It does mean we'll need to add an `after()` block into any spec that results in a DB connection being made. But that should be simple enough.

Completing this example has also highlighted that we'll probably want to extend `assert` in some way, to avoid 'messy duplication' in our tests.

But for now, here is our first working model node-test!