Add example of simple mocking
We've picked `CheckBusyBillRunsService`, a service in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system), as our first candidate to demonstrate mocking.

It relies in our `db/db.js` module to make a raw query to the DB. In **water-abstraction-system** we're able to mock out the call to `select()` and therefore easily provide the data `CheckBusyBillRunsService` needs to exercise its logic.

We wanted to try and replicate this simple example in node-test runner, and we have!