Add option to run tests without cleaning
The test suite automatically cleans and seeds the test DB before any tests are run. When running the full suite this ensures we are running with a 'known' environment and avoids flaky tests.

But when working on a specific feature, you know where the changes are being made and might want to avoid the extra time the clean and seed takes before running the tests.

This change adds the option to skip as a script to `package.json`.

The call is the same as the current one for `test`. The fact it isn't called `test:` means [npm](https://www.npmjs.com/) won't call `pretest:` before it is run. Simples!