Update CI test script to use multiple reporters
We thought we only had the option to set one reporter, and for CI we need to code coverage in `lcov` format to support SonarCloud.

But we have since learnt we can specify multiple reporters. This change means when we look at the job in CI we can see the same output we get locally whilst still generating code coverage for SonarCloud.