Add a standalone docker-based environment
We know there are improvements we can take advantage of if we move to Node v22 when it comes to working with node-test. We thought we just needed to blip our current **wal-dev-environment** to Node v22 and continue working on this project in there to try the improvements out.

However, an issue with node-gyp in [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) is preventing the environment building all apps under Node v22. So till that is resolved, we've been forced to create a standalone environment to run this project in that uses Node v22.

This change adds support for running the project in its own docker environment. The files have been taken from our main water-abstraction local environment project so the build process and resulting environment is similar.