Rename specs folder back to test
Because Node v20 was limited in what could be done to control what it considered to be a 'test' we had to rename our `test/` folder as `specs/`. Node v22 now supports glob patterns so we can specify we _only_ want it to consider files named `*.test.js` as tests, no matter where they are.

This means after renaming our `specs/` folder as `test/` it will stop treating the support files in there as tests.