Maker test coverage more accurate
One of the reasons for pushing the service to use Node v22 is to take advantage of the additional support it has for the node test runner, one of which being the ability to configure what files code coverage should be reported for. Admittedly, it is still experimental, so how it works may change in the future. But we foresee the functionality only improving and expanding.

This change updates the args we pass on the command line (again, hopefully they'll be some way to move this to a shared config in the future!) to exclude those files we are not interested in reporting coverage on.