Fix PM2 erroring when docker environment starts
When we added prettier and the updated ESLint config we automatically updated the pm2 config file to ESM as well.

But there was a reason it still used CommonJS syntax and has a `.cjs` extension: PM2 can't handle the config file being written as ESM!

We did know this, we just forgot. This puts back the file (with the changes since then) and updates our ESLint config to stop shouting at us about it!