Add example test for base.request.js
The `base.request.js` module in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) is a great candidate for rewriting its tests using the node-test runner.

In the module itself, we can bring in [Got](https://github.com/sindresorhus/got) using a standard import. In the tests, we bring in and [Nock](https://github.com/nock/nock) for the first time and confirm that works as before.

We also deal with mocking our `GlobalNotifier` for the first time (we've not even added it yet!) and make assertions against it.

The key things we're hoping to get to the bottom of though is

- how would we alter the value of one of our config files, as we can do with `Sinon.replace()`
- will we need to use the [timeout option](https://nodejs.org/api/test.html#testname-options-fn) like we do in Lab for our timeout tests

Plus, as while we are working on a new test, we give the updated `--test-only` functionality a run for its money!