Add support for test only
We've confirmed we can now place our `.only()` anywhere in the test file, and it will only run it, or those in its block (we don't have to place `.only()` at every level any more).

However, currently you do still have to add the flag on the command line to node in order for it to look for `.only()` tests. If you add the command line flag and don't put `.only()` anywhere, nothing gets run.

This is why we have added an additional VSCode task to allow us to run tests in 'only' mode.