Blocked replacing the property
Not added here, but we have tried to add the first `describe('because request timed out', () => {` test. However, we are currently stumped trying to find an alternate to `Sinon.replace()`.

What we can confirm is unlike Lab, node-test has no timeout on each individual test. You can specify one either at test or at global level. But if you don't then there is no timeout.

So, we can get the test to work if we tell Nock to delay for 6000ms (our request timeout is 5000ms). But obviously we don't want to really delay our test suite for 6 seconds. Hence we are currently blocked progressing these tests.